refactor: move warnings components
