+++ frontend/src/lib/components/warnings/TransactionRateWarningIcon.svelte
