+++ frontend/src/lib/components/warnings/Warnings.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import TransactionRateWarning from "$lib/components/metrics/TransactionRateWarning.svelte";
+  import TransactionRateWarning from "$lib/components/warnings/TransactionRateWarning.svelte";
   import { ENABLE_METRICS } from "$lib/constants/mockable.constants";
   import Metrics from "$lib/components/metrics/Metrics.svelte";
   import { Toasts } from "@dfinity/gix-components";
