+++ frontend/src/tests/lib/components/metrics/WarningsTest.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import Metrics from "$lib/components/metrics/Metrics.svelte";
-  import TransactionRateWarning from "$lib/components/metrics/TransactionRateWarning.svelte";
+  import TransactionRateWarning from "$lib/components/warnings/TransactionRateWarning.svelte";
   import { Toasts } from "@dfinity/gix-components";
 </script>
